Refactored DownloadDialogue to have common base class. - Merged together all DownloadOptions classes. - Eliminated some unnecessary import statements.
