+++ src/extensions/uv-pdf-extension/DownloadDialogue.ts
@@ -1,17 +1,7 @@
-import BaseCommands = require("../../modules/uv-shared-module/Commands");
-import Dialogue = require("../../modules/uv-shared-module/Dialogue");
-import DownloadOption = require("./DownloadOption");
-import IPDFExtension = require("./IPDFExtension");
-import IPDFProvider = require("./IPDFProvider");
-import RenderingFormat = require("../../modules/uv-shared-module/RenderingFormat");
-import ServiceProfile = require("../../modules/uv-shared-module/ServiceProfile");
-import Shell = require("../../modules/uv-shared-module/Shell");
+import BaseDownloadDialogue = require("../../modules/uv-dialogues-module/DownloadDialogue");
+import DownloadOption = require("../../modules/uv-dialogues-module/DownloadOption");
 
-class DownloadDialogue extends Dialogue {
-
-    $downloadOptions: JQuery;
-    $noneAvailable: JQuery;
-    $title: JQuery;
+class DownloadDialogue extends BaseDownloadDialogue {
 
     constructor($element: JQuery) {
         super($element);
@@ -22,27 +12,6 @@ class DownloadDialogue extends Dialogue {
         this.setConfig('downloadDialogue');
 
         super.create();
-
-        $.subscribe(BaseCommands.SHOW_DOWNLOAD_DIALOGUE, (e, params) => {
-            this.open();
-        });
-
-        $.subscribe(BaseCommands.HIDE_DOWNLOAD_DIALOGUE, (e) => {
-            this.close();
-        });
-
-        // create ui.
-        this.$title = $('<h1>' + this.content.title + '</h1>');
-        this.$content.append(this.$title);
-
-        this.$noneAvailable = $('<div class="noneAvailable">' + this.content.noneAvailable + '</div>');
-        this.$content.append(this.$noneAvailable);
-
-        this.$downloadOptions = $('<ol class="options"></ol>');
-        this.$content.append(this.$downloadOptions);
-
-        // hide
-        this.$element.hide();
     }
 
     open() {
@@ -70,26 +39,11 @@ class DownloadDialogue extends Dialogue {
         this.resize();
     }
 
-    addEntireFileDownloadOption(fileUri: string): void{
-        this.$downloadOptions.append('<li><a href="' + fileUri + '" target="_blank" download>' + String.format(this.content.entireFileAsOriginal, this.getFileExtension(fileUri)) + '</li>');
-    }
-
-    getFileExtension(fileUri: string): string{
-        return fileUri.split('.').pop();
-    }
-
     isDownloadOptionAvailable(option: DownloadOption): boolean {
         switch (option){
             case DownloadOption.entireFileAsOriginal:
                 return true;
         }
-    }
-
-    resize(): void {
-
-        this.$element.css({
-            'top': this.extension.height() - this.$element.outerHeight(true)
-        });
     }
 }
