--- src/extensions/uv-pdf-extension/DownloadOption.ts
@@ -1,12 +1,0 @@
-class DownloadOption {
-    static entireFileAsOriginal = new DownloadOption("entireFileAsOriginal");
-
-    constructor(public value: string) {
-    }
-
-    toString() {
-        return this.value;
-    }
-}
-
-export = DownloadOption;
