+++ src/modules/uv-dialogues-module/DownloadOption.ts
@@ -3,6 +3,7 @@ class DownloadOption {
     static dynamicCanvasRenderings = new DownloadOption("dynamicCanvasRenderings");
     static dynamicImageRenderings = new DownloadOption("dynamicImageRenderings");
     static dynamicSequenceRenderings = new DownloadOption("dynamicSequenceRenderings");
+    static entireFileAsOriginal = new DownloadOption("entireFileAsOriginal");
     static wholeImageHighRes = new DownloadOption("wholeImageHighRes");
     static wholeImageLowResAsJpg = new DownloadOption("wholeImageLowResAsJpg");
